requirements.txt
certifi
openpyxl
pandas
pyinstaller
pyyaml
